Revert "lineage-sdk self-removing prefs: move reap from onBindViewHolder to onAttached"
This reverts commit a89a6e11cfbe1163892b7f2fafd5075661112bb2.

While it served its original purpose, this was causing bad side-effects
when using our constraints (verified at least with requiresConfig and
requiresAction). Moving the reap to onAttached was rendering self-removing
prefs useless because if the view isn't loaded yet, the preference
(which also doesn't yet exists) can't really be hidden.

Change-Id: Ic61fe4b9979412d89c55ea1543c494e69bf6eab7